# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Loopsy API
  version: "1.0.0"
  description: >
    Loopsy — the AI-native crochet studio. API-only Next.js backend. Every
    front door (text, photo, canvas) emits a Design Spec; the deterministic
    engine computes and verifies all stitch counts. Auth is cookie-session
    based; AI generation is metered per plan. Generated from the verified route
    inventory (docs/api/01-endpoint-catalog.md).
servers:
  - url: https://loopsy-hazel.vercel.app
    description: Production (SPA origin; /api/* is rewritten to the Railway backend)
  - url: http://localhost:5173
    description: Local (Vite dev proxy → backend on :3000)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: loopsy_session
      description: Opaque session token set on signup/login; httpOnly, SameSite=Lax, Secure in prod.
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        details: { type: string }
      required: [error]
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        emailVerified: { type: integer, enum: [0, 1] }
        subscription:
          type: object
          properties:
            plan: { type: string, enum: [free, maker_pro, creator] }
            status: { type: string }
...
    Design:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        spec: { type: object, additionalProperties: true }
        patternId: { type: [string, "null"] }
  responses:
    Unauthorized:
      description: Not signed in.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    BadRequest:
      description: Validation error.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    RateLimited:
      description: Plan limit reached (code RATE_LIMIT_EXCEEDED) or throttled.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
paths:
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Create an account (sends an email-verification link)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string, minLength: 1 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Created; sets the session cookie.
          content: { application/json: { schema: { type: object, properties: { user: { $ref: "#/components/schemas/User" } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { description: Cross-site request blocked (CSRF). }
        "409": { description: Email already registered. }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Set a new password using a reset token
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [token, password], properties: { token: { type: string }, password: { type: string, minLength: 8 } } } } }
      responses:
        "200": { description: Password updated. }
        "400": { description: Invalid/expired token or weak password. }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/designs:
    get:
      tags: [Designs]
      summary: List the caller's designs
      security: [{ cookieAuth: [] }]
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Design" } } } } }, "401": { $ref: "#/components/responses/Unauthorized" } }
    post:
      tags: [Designs]
      summary: Save a canvas design
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [spec], properties: { name: { type: string }, spec: { type: object, additionalProperties: true } } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Design" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...